refactored and merged with master
